Add initial NetworkLogger.URLSessionProxy implementation
